+++ packages/cli/src/utilities/index.ts
@@ -6,6 +6,7 @@ export * from './error';
 export * from './feature';
 export {merge} from './merge';
 export {componentName, validComponentName} from './react';
+export {formatFile} from './format';
 
 const DEFAULT_SUBCOMMANDS = {
   create: 'app',
@@ -26,16 +27,4 @@ export function parseCliArguments(rawInputs?: string[]) {
     mode: debug ? 'debug' : 'default',
     command: [command, subcommand].join('/'),
   };
-}
-
-export function formatFile(file: string) {
-  const match = file.match(/^[^\S\n]*(?=\S)/gm);
-  const indent = match && Math.min(...match.map((el) => el.length));
-
-  if (indent) {
-    const regexp = new RegExp(`^.{${indent}}`, 'gm');
-    return file.replace(regexp, '').trim() + '\n';
-  }
-
-  return file.trim() + '\n';
 }