+++ packages/eslint-plugin/src/rules/no-effects-in-server-components/examples/valid.example.tsx
@@ -1,0 +1,11 @@
+// Examples of **correct** code for this rule:
+
+// MyClientComponent.client.jsx
+import {useEffect} from 'react';
+
+function MyClientComponent() {
+  useEffect(() => {
+    // in client components, this code will execute as expected
+  });
+  return null;
+}