+++ packages/hydrogen/CHANGELOG.md
@@ -7,6 +7,25 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+- fix: default to `retry: false` in `useQuery`
+- fix: starter template media gallery error when handling videos
+- fix: run graphiql middleware before vite, fixing graphiql
+- feat: allow custom Model3D poster
+
+## 0.6.4 - 2021-11-11
+
+- fix: let Vite handle public assets in development
+- fix: new lines in hydration request break JSON.parse
+- fix(#201): normalize POSIX separators to support windows
+- fix: scroll to top on app first load
+
+## 0.6.3 - 2021-11-10
+
+- fix: add trailing slash to user components glob
+
+## 0.6.2 - 2021-11-10
+
+- fix: remove CartProvider from BuyNowButton
 - fix: reading property of null for component props
 - fix: transform deeply-imported client components
 - fix: duplicated files and contexts in browser